backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1
python-multipart==0.0.6
httpx==0.25.1

# AI Service SDKs
openai==1.3.7
anthropic==0.39.0
google-generativeai==0.3.0
azure-cognitiveservices-speech==1.34.0
google-cloud-speech==2.22.0
elevenlabs==0.2.27